Добавляет наложение эффекта
