+++ js/form.js
@@ -1,13 +1,7 @@
 import {isEscapeKey} from './util.js';
-import {onPhotoScaleUpButtonClick, onPhotoScaleDownButtonClick} from './validate-form.js';
-
-const MAX_HASHTAGS_COUNTS = 5;
-const MAX_COMMENT_LENGTH = 140;
-const REGEX_HASHTAGS = /^#[a-zа-яё0-9]{1,19}$/i;
-const INVALID_HASHTAG_FORMAT = 'Хэш-тег должен начинаться с решетки, содержать только буквы и цифры, иметь длину не более 20 символов';
-const INVALID_HASHTAG_COUNT = 'Количество хэш-тегов не больше пяти';
-const INVALID_HASHTAG_REPEAT = 'Хэш-теги не должны повторяться';
-const INVALID_COMMENT_LENGTH = 'Длина комментария не может составлять больше 140 символов';
+import {setScale, resetScale} from './scale.js';
+import {setEffectSlider} from './effects.js';
+import {isValid, resetValidator} from './validator.js';
 
 const uploadField = document.querySelector('.img-upload__input');
 const uploadOverlay = document.querySelector('.img-upload__overlay');
@@ -15,48 +9,9 @@ const uploadCancelButton = document.querySelector('.img-upload__cancel');
 const uploadForm = document.querySelector('.img-upload__form');
 const hashtagsInput = uploadForm.querySelector('.text__hashtags');
 const commentInput = uploadForm.querySelector('.text__description');
-const photoScaleDownButton = document.querySelector('.scale__control--smaller');
-const photoScaleUpButton = document.querySelector('.scale__control--bigger');
-
-const erorrConfig = {
-  classTo: 'img-upload__field-wrapper',
-  errorTextParent: 'img-upload__field-wrapper'
-};
-
-const pristine = new Pristine(uploadForm, erorrConfig);
-
-const normalizeHashtags = (tags) => tags
-  .trim()
-  .toLowerCase()
-  .split(' ')
-  .filter((hashtag) => hashtag.length > 0);
-
-const validateHashtagsCount = (value) => {
-  const hashtags = normalizeHashtags(value);
-  return hashtags.length <= MAX_HASHTAGS_COUNTS;
-};
-
-const validateHashtagsFormat = (value) => {
-  const hashtags = normalizeHashtags(value);
-  return hashtags.every((hashtag) => REGEX_HASHTAGS.test(hashtag));
-};
-
-const validateHashtagsDublicate = (value) => {
-  const hashtags = normalizeHashtags(value);
-  return hashtags.length === new Set(hashtags).size;
-};
-
-const validateCommentLength = (value) => value.length <= MAX_COMMENT_LENGTH;
-
-pristine.addValidator(hashtagsInput, validateHashtagsCount, INVALID_HASHTAG_COUNT);
-pristine.addValidator(hashtagsInput, validateHashtagsFormat, INVALID_HASHTAG_FORMAT);
-pristine.addValidator(hashtagsInput, validateHashtagsDublicate, INVALID_HASHTAG_REPEAT);
-pristine.addValidator(commentInput, validateCommentLength, INVALID_COMMENT_LENGTH);
 
 const onSubmitButtonClick = (evt) => {
   evt.preventDefault();
-
-  const isValid = pristine.validate();
 
   if(isValid) {
     uploadForm.submit();
@@ -65,14 +20,15 @@ const onSubmitButtonClick = (evt) => {
 
 const closeForm = () => {
   uploadForm.reset();
-  pristine.reset();
+  resetValidator();
+  resetScale();
   document.removeEventListener('keydown', onDocumentKeydown);
   uploadOverlay.classList.add('hidden');
   document.body.classList.remove('.modal-open');
   uploadForm.removeEventListener('submit', onSubmitButtonClick);
   uploadCancelButton.removeEventListener('click', onButtonCloseUploadForm);
-  photoScaleDownButton.removeEventListener('click', onPhotoScaleDownButtonClick);
-  photoScaleUpButton.removeEventListener('click', onPhotoScaleUpButtonClick);
+  //photoScaleDownButton.removeEventListener('click', onPhotoScaleDownButtonClick);
+  //photoScaleUpButton.removeEventListener('click', onPhotoScaleUpButtonClick);
 };
 
 const openForm = () => {
@@ -81,8 +37,8 @@ const openForm = () => {
   document.addEventListener('keydown', onDocumentKeydown);
   uploadForm.addEventListener('submit', onSubmitButtonClick);
   uploadCancelButton.addEventListener('click', onButtonCloseUploadForm);
-  photoScaleDownButton.addEventListener('click', onPhotoScaleDownButtonClick);
-  photoScaleUpButton.addEventListener('click', onPhotoScaleUpButtonClick);
+  setEffectSlider();
+  setScale();
 };
 
 function onButtonCloseUploadForm () {
