+++ js/main.js
@@ -1,6 +1,6 @@
 import {createPhotos} from './data.js';
 import {renderGallery} from './gallery.js';
 import './form.js';
-import './effects.js';
+//import './effects.js';
 
 renderGallery(createPhotos());
